% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fqc.R
\name{aggregate_fqc}
\alias{aggregate_fqc}
\alias{module_fqc}
\alias{stats_fqc}
\title{Aggregate and summarise FastQC reports for multiple samples}
\usage{
aggregate_fqc(object)

module_fqc(object)

stats_fqc(object)
}
\arguments{
\item{object}{A fqcRes.multi object produced by `read_multifastqc()`.}
}
\value{
\itemize{
    \item \strong{aggregate_fqc()} returns a list containing two data-frames.
    The first data-frame has the following columns:
         \itemize{
         \item sample: sample names
         \item tot.seq: total sequences (i.e.: the number of reads)
         \item low.qual: the number of sequences flagged as poor quality
         \item seq.len: sequence length
         \item pct.gc: \% of GC content
         \item pct.dup: \% of duplicate reads
         }
     In the second datframe, each row is a sample and the columns are the
     FASTQC modules.  Each cell indicates the outcome (pass, warn or fail)
     for that module by each sample.
   \item \strong{module_fqc()}: Generates a list of summarising the
   module results.
         \itemize{
         \item The first list item is a dataframe, in which each row is
   a sample and the columns are the FASTQC modules.  Each cell indicates the
   outcome (pass, warn or fail) for that module by each sample.
         \item The second item summarises the module results per module,
         indicating for each module, how many samples passed, warned or
         failed.  In addition, individual summaries for warned and failed are
         provided.
         \item The third item summrises the module results per sample,
         indicating for each sample, how many modules passed, warned or
         failed.  In addition, individual summaries for warned and failed are
         provided.
         }
   \item \strong{stats_fqc()}: returns a data frame containing
   basic metrics of the fastqc reports.  Columns are: sample, tot.seq,
   low.qual.seq, seq.len, pct.gc, and pct.dup.
   }
}
\description{
Aggregate and summarise multiple FastQC reports.
}
