% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rna.R
\name{align_rna}
\alias{align_rna}
\title{Alignment of RNA-seq reads}
\usage{
align_rna(threads = 1, plot = TRUE, out.dir = ".", bigwig = FALSE,
  hisat2 = "hisat2", samtools = "samtools", sambamba = "sambamba",
  idx = NULL, reads.dir = NULL, reads1 = NULL, reads2 = NULL,
  fastq = TRUE, fasta = FALSE, softClipPenalty = NULL,
  noSoftClip = FALSE, knownSplice = NULL, strand = NULL, tmo = FALSE,
  secondary = FALSE, maxAlign = NULL, nomixed = FALSE,
  nodiscordant = FALSE, rgid = NULL, quiet = FALSE,
  non_deterministic = TRUE, memory = "1G", remove.mitochondrial = "ChrM",
  hash_table = 262144, overflow_size = 2e+05, io_buffer = 128,
  scale = 2e+07)
}
\arguments{
\item{threads}{A positive integer specifying the number of sorting and
compression threads}

\item{plot}{Logical.  If TRUE, a PNG summarising the alignment statistics is
produced}

\item{out.dir}{Character string.  Directory to save results in.  If the
directory does not exist, it will be created}

\item{bigwig}{Logical indicating whether bigwig files should be created.
This is memory and time-consuming.}

\item{hisat2}{The path to hisat2 (if not in executable path).}

\item{samtools}{The path to samtools (if not in executable path).}

\item{sambamba}{The path to sambamba (if not in executable path).}

\item{idx}{The basename of the index for the reference genome. The basename
is the name of any of the index files up to but not including the final
.1.ht2, etc.}

\item{reads.dir}{Character string.  Directory containing the raw reads.  If
this is specified, then reads1 and reads2 should be set as NULL.}

\item{reads1}{Character vector of mate1 reads.  If specified, then reads.dir
must be NULL.}

\item{reads2}{Character vector of mate2 reads.  If specified, then reads.dir
must be NULL.  Must be the same length as mate1.  If single-end sequencing,
then should be left as NULL.}

\item{fastq}{Logical indicating if reads are FASTQ files.}

\item{fasta}{Logical indicating if reads are FASTA files.}

\item{softClipPenalty}{Sets the maximum (MX) and minimum (MN) penalties for
soft-clipping per base, both integers.  Must be given in the format "MX,MN".}

\item{noSoftClip}{Logical indicating whether to disallow soft-clipping.}

\item{knownSplice}{Path to text file containing known splice sites.}

\item{strand}{Specify strand-specific information.  Default is unstranded.}

\item{tmo}{Logical indicating whether to report only those reads aligning to
known transcripts.}

\item{secondary}{Logical indicating whether to report secondary alignments.}

\item{maxAlign}{Integer indicating the maximum number of distinct primary
alignments to search for each read.}

\item{nomixed}{By default, when hisat2 cannot find a concordant or discordant
alignment for a pair, it then tries to find alignments for the individual
mates. If TRUE, this option disables that behavior.}

\item{nodiscordant}{By default, hisat2 looks for discordant alignments if it
cannot find any concordant alignments.  If true, this option disables that
behavior.}

\item{rgid}{Character string, to which the read group ID is set.}

\item{quiet}{If TRUE, print nothing except alignments and serious errors.}

\item{non_deterministic}{When set to TRUE, HISAT2 re-initializes its
pseudo-random generator for each read using the current time.#' @param rgid}

\item{memory}{String specifying maximum memory per thread; suffix K/M/G
recognized.}

\item{remove.mitochondrial}{Character string.  If set, this will remove reads
mapping to the mitochondrial genome.  The string should match the reference
name for the mitochindrial genome in the alignment file.  Examples include
"ChrM", "M" and "MT".}

\item{hash_table}{Size of hash table for finding read pairs (default is
262144 reads); will be rounded down to the nearest power of two.  For best
performance should be > (average coverage) * (insert size).}

\item{overflow_size}{Size of the overflow list where reads, thrown out of
the hash table, get a second chance to meet their pairs (default is
200000 reads); increasing the size reduces the number of temporary files
created.}

\item{io_buffer}{Controls sizes of the two buffers (in MB) used for reading
and writing BAM during the second pass (default is 128).}

\item{scale}{The number of reads to scale to.  If NULL no scaling performed.}
}
\value{
Raw and filtered BAM files +/- bigwig files
}
\description{
Alignment of RNA-seq reads
}
\examples{
\dontrun{
align_rna(threads = 2, plot = TRUE, out.dir = ".", bigwig = FALSE,
          hisat2 = "hisat2", samtools = "samtools", sambamba = "sambamba",
          idx = "../prana/data-raw/index/UCSC.hg19",
          reads.dir = "../prana/data-raw/seqFiles", fastq = TRUE)
}
}
