\name{annotation}
\alias{annotation}
\docType{data}
\title{
SAF file for human hg19 genome
}
\description{
Human hg19 genome in SAF format
}
\usage{data("annotation")}
\format{
  A data frame with 963553 observations on the following 5 variables.
  \describe{
    \item{\code{GeneID}}{gene name}
    \item{\code{Chr}}{chromsome}
    \item{\code{Start}}{start position}
    \item{\code{End}}{end position}
    \item{\code{Strand}}{strand location, either positive or negative strand}
  }
}
\examples{
data(annotation)
}

\keyword{datasets, annotation, hg19, genome}
