% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fqreport.R
\name{create_fqreport}
\alias{create_fqreport}
\title{Create a QC report}
\usage{
create_fqreport(fqc = NULL, outdir = ".", experiment = NULL,
  author = NULL, preview = TRUE)
}
\arguments{
\item{fqc}{Either a single zipped FASTQC file or a directory containing such
files.}

\item{outdir}{The directory to which to write out the results.  The default
is the current working directory.}

\item{experiment}{Brief experiment description}

\item{author}{Author name}

\item{preview}{Logical indicating whether to open the HTML file after it is
created.}
}
\value{
An HTML file (or files) containing FASTQC reports.
}
\description{
Create an HTML file containing FastQC reports of one or multiple
 files. Inputs can be either a directory containing multiple FastQC zipped
 files or a single FastQC file.
}
