% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hmp.R
\name{create_hmp}
\alias{create_hmp}
\title{Heatmap of most differential genes.}
\usage{
create_hmp(results, n = 100, main = "Heatmap of top DEGs",
  show_row_names = FALSE)
}
\arguments{
\item{results}{Results slot from object created by \code{\link{edger_pipe}},
\code{\link{limma_pipe}}, or \code{\link{deseq2_pipe}}.}

\item{n}{Topmost number of genes to plot.}

\item{main}{Title of plot.}

\item{show_row_names}{Boolean indicating whether to show gene names in the
heatmap.}
}
\value{
Heatmap in PNG format.
}
\description{
Heatmap showing centred and normalised expression across each sample
for the most differentially-expressed genes in each contrast.
}
\examples{
\dontrun{
load(system.file("data", "er.rda", package = "dealr"))
create_hmp(er$results)
}
}
