% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MAplot.R
\name{create_maplot}
\alias{create_maplot}
\title{Create MA plot}
\usage{
create_maplot(res, analysis = "limma")
}
\arguments{
\item{res}{The output of either \code{\link{run_edger}},
\code{\link{run_limma}}, or \code{\link{run_deseq2}}.}

\item{analysis}{The analysis method - one of "edger", "voom", or "deseq2".}
}
\value{
MA plot in PNG and PDF formats.
}
\description{
Function to create Mean-Average (Bland-Altmann) plots showing the log-fold
changes in gene expression versus the average expression of that gene.
}
\examples{
\dontrun{
create_maplot(results, analysis = "edger")
}
}
