% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mds.R
\name{create_mds}
\alias{create_mds}
\title{Create multi-dimensional scaling plot from DGE object}
\usage{
create_mds(dge = dge, main = "", dim.plot = c(1, 2), n = min(500,
  nrow(dge$counts)))
}
\arguments{
\item{dge}{DGE object produced by \code{\link[edgeR]{DGEList}}.}

\item{main}{Plot title.}

\item{dim.plot}{Dimensions to plot - default are dimensions 1 and 2.}

\item{n}{Number of pairwise genes to consider when plotting MDS.  Default is
500.}
}
\value{
MDS plot.
}
\description{
Function to create a MDS plot from a DGE object.
}
\examples{
\dontrun{
create_mds(dge)
}
}
