% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pca.R
\name{create_pca}
\alias{create_pca}
\title{Create principal component analysis plot from DDS object}
\usage{
create_pca(dds)
}
\arguments{
\item{dds}{DESeq2 dds object created by
\code{\link[DESeq2]{DESeqDataSetFromMatrix}}.}
}
\value{
PCA plot in PNG and PDF formats.
}
\description{
Function to create a PCA plot from a DDS object
}
\examples{
\dontrun{
load(system.file("data", "ddsMat.rda", package = "dealr"))
create_pca(ddsMat)
}
}
