% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_norm.R
\name{deseq2_norm}
\alias{deseq2_norm}
\title{Count normalisation using DESeq2}
\usage{
deseq2_norm(dds = dds)
}
\arguments{
\item{dds}{DESeq2 dds object created by
\code{\link[DESeq2]{DESeqDataSetFromMatrix}}.}
}
\value{
Boxplots showing the effect of normalisation in PNG and PDF format.
}
\description{
Function to show the effect of the count normalisation that will be
performed as part of the DESeq2 analysis.
}
\examples{
\dontrun{
load(system.file("data", "ddsMat.rda", package = "dealr"))
deseq2_norm(dds)
}
}
