% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_norm.R
\name{edger_norm}
\alias{edger_norm}
\title{Count normalisation using edgeR}
\usage{
edger_norm(dge = NULL, norm.method = "TMM")
}
\arguments{
\item{dge}{DGE object.}

\item{norm.method}{Normalisation method to use - may be "TMM", "RLE",
"upperquartile" or "none".}
}
\value{
Normalised DGE object.
}
\description{
Function to perform count normalisation prior to differential analysis,
with generation of before and after plots.
}
\examples{
\dontrun{
dge <- edger_norm(dge, norm.method = "TMM")
}
}
