\name{grammar.symbols}
\alias{grammar.symbols}
\title{
	Read the grammar.output file generated by bison
}
\description{
	When bison is used with the \code{-v} switch, the 
	\code{grammar.output} file is generated. This function extracts
	symbols, terminal and non-terminal, from the file
	and structure them in a data frame
}
\usage{
grammar.symbols()
}
\value{
	A data frame. See \link{symbols}
}
\author{
	Romain Francois <romain@r-enthusiasts.com>
}
\examples{
\dontrun{
# the bison generated grammar output
system.file( "grammar", "gram.output", package = "parser" )

grammar.symbols()
}
}
\keyword{ manip }

