\docType{data}
\name{grammar_symbols}
\alias{grammar_symbols}
\title{Parser Grammar Symbols}
\format{A data frame with \code{desc}, the description;
 \code{token}, the number identifier of the class;
 and \code{terminal} if the class constitutes terminal nodes.}
\description{
  A data table with the symbols used in the parser.
}
\keyword{datasets}

