% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_motifs.R
\name{homer_known}
\alias{homer_known}
\alias{homer_denovo}
\title{Parse HOMER motif files}
\usage{
homer_known(motif_dir = "motifs")

homer_denovo(motif_dir = "motifs")
}
\arguments{
\item{motif_dir}{Directory in which the motif files created by HOMER
are located}
}
\value{
An R object with the known motifs in an easily readable format.

An R object with the \emph{de novo} motifs in an easily readable format.
}
\description{
\code{homer_known} is a function to parse the 'known' motifs
  identified by HOMER.

\code{homer_denovo} is a function to parse the 'denovo' motifs identified by
  HOMER.
}
\examples{
\dontrun{
motifdir <- system.file("extdata", "motifs.tar.gz", package = "dealr")
extract <- sprintf("tar -xvzf \%s", motifdir)
system(extract)
motifs <- list.files("motifs")
hk <- homer_known("motifs")
hd <- homer_denovo("motifs")
compress <- sprintf("tar -cvzf \%s \%s", motifs, motifdir)
system(compress)
unlink("motifs", recursive = TRUE)
}
}
