% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{iclean}
\alias{iclean}
\title{Clean document texts}
\usage{
iclean(text, lowercase = TRUE, replacechars = c(j = "i", v = "u"),
  removeangbracketed = TRUE, removeparensed = TRUE,
  onlycharacter = TRUE, removepunctuation = FALSE,
  removestopwords = FALSE, language = "lat", customstopwords = NULL,
  removenumbers = FALSE, removeextraspaces = TRUE)
}
\arguments{
\item{text}{vector of character strings}

\item{lowercase}{logical: whether to convert to lower case}

\item{replacechars}{names vector showing how to substitute characters to normalize spelling}

\item{removeangbracketed}{logical: whether to remove text between angle brackets}

\item{removeparensed}{logical: whether to remove text in parentheses}

\item{onlycharacter}{logical: whether to strip all but alphanumeric characters}

\item{removepunctuation}{logical: whether to remove punctuation (matters only if \code{onlycharacter=FALSE})}

\item{removestopwords}{logical: whether to remove stopwords. CURRENTLY HAS NO EFFECT.}

\item{language}{logical: what language stopwords to use. CURRENTLY HAS NO EFFECT.}

\item{customstopwords}{a vector of stopwords to remove.}

\item{removenumbers}{logical, whether to remove numeric characters.}

\item{removeextraspaces}{logical, whether to trim extra spaces.}
}
\value{
a vector of character strings, the clean version of \code{text}.
}
\description{
Clean document texts
}
