% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ilistForms}
\alias{ilistForms}
\title{List word forms found by a logical condition}
\usage{
ilistForms(parsed, where = "x$lemma==\\"UNKNOWN\\"")
}
\arguments{
\item{parsed}{a \code{parsed} object}

\item{where}{a character string containing a logical expression. The analysis list should be named \code{x}. E.g., if looking for all analyses with lemma "dog", the argument should be 'x$lemma=="dog"'}
}
\value{
a sorted string vector of all distinct word forms found. If the argument \code{where} is an empty string it is interpreted as no condition, and the output will contain all word forms present in \code{parsed}.
}
\description{
List all word forms from a \code{parsed} object that sarisfy a logical condition
}
