% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{iparseWord}
\alias{iparseWord}
\title{Parse a word}
\usage{
iparseWord(w,
  URLPrefix = "http://www.perseus.tufts.edu/hopper/xmlmorph?lang=la&lookup=")
}
\arguments{
\item{w}{the word}

\item{URLPrefix}{the url of the XML parsing results for each word is constructed by prefixing URLPrefix to the word. The default value will query Perseus Latin word tool database.}
}
\value{
A list of the same structure as a \code{word} element in a \code{parsed} data object, such as produced by the function \code{iparse}.
}
\description{
Parse a single word by making a call to a lookup parsing service
}
