% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{iremoveLookupErrors}
\alias{iremoveLookupErrors}
\title{Remove lookup errors}
\usage{
iremoveLookupErrors(parsed,
  URLPrefix = "http://www.perseus.tufts.edu/hopper/xmlmorph?lang=la&lookup=")
}
\arguments{
\item{parsed}{an object as produced by \code{iparse}.}

\item{URLPrefix}{the url of the XML parsing results for each word is constructed by prefixing URLPrefix to the word. The default value will query Perseus Latin word tool database.}
}
\value{
the updated version of the object \code{parsed}.
}
\description{
Takes a data object \code{parsed}, such as produced by the function \code{iparse}, finds all lookup errors (instances where the call to the parsing database failed, resulting in \code{lemma=’ERROR’}), makes the call for these instances and updates the data object.
}
