% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{iwhich}
\alias{iwhich}
\title{Query a \code{parsed} object}
\usage{
iwhich(parsed, where = "x$lemma==\\"UNKNOWN\\"", absent = FALSE)
}
\arguments{
\item{parsed}{a data object, such as produced by the function \code{iparse}}

\item{where}{a character string containing a logical expression. The analysis list should be named \code{x}. E.g., if looking for all analyses with lemma "dog", the argument should be 'x$lemma=="dog"'}

\item{absent}{a logical, controlling what happens when the queried field is absent: should it be treated as satisfying (TRUE) or not satisfying (FALSE) the query.}
}
\value{
a matrix with columns \code{document}, \code{word}, \code{analysis}. Each row contains the number of the document, word, analysis where the search result was found. If no results were found the output is NULL.
}
\description{
This function is to perform queries on the data object \code{parsed} , such as produced by \code{iparse}.
}
