% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_pipe.R
\name{limma_pipe}
\alias{limma_pipe}
\title{Pipeline for limma+voom analysis}
\usage{
limma_pipe(se = NULL, design = NULL, contrast = NULL,
  block.column = NULL, norm.method = "TMM", adjust.method = "BH",
  p.value = 0.05)
}
\arguments{
\item{se}{SummarizedExperiment with an assay slot named counts_fil containing
the filtered counts and colData slot containing sample metadata.}

\item{design}{A design matrix specifying the experimental design.  If NULL,
then a design matrix will be created using the values for contrast.levels
and block.levels.}

\item{contrast}{A matrix of contrasts (as created by \code{makeContrasts} in
the \code{limma} package specifying the comparisons to perform).  If NULL,
then will be created from the design matrix.}

\item{block.column}{The column in the samples metadata dataframe specifying
the block/additive effect column.  Ignored if design is not NULL.}

\item{norm.method}{Normalisation method to use - may be "TMM", "RLE",
"upperquartile" or "none".}

\item{adjust.method}{Method to be used for adjustment of nominal p-values.
May be one of "BH", "bonferroni", "holm", "hochberg", "hommel", "BY".}

\item{p.value}{Value between 0 and 1.  Adjusted p-value for the differential
expression analysis.}
}
\value{
Results of limma+voom differential analysis.
}
\description{
Function to perform entire limma+voom analysis using voom with quality
weights, and beginning with a SummarizedExperiment and ending with list of
differential genes and generation of diagnostic plots along the way.
}
