% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_samples.R
\name{load_samples}
\alias{load_samples}
\title{Load sample metadata file}
\usage{
load_samples(sample.path = NULL, contrast.column = NULL,
  block.column = NULL, contrast.levels)
}
\arguments{
\item{sample.path}{Character string.  Full path to sample metadata file.}

\item{contrast.column}{Chacter string.  Column in sample metadata file
that contains the contrasts of interest.}

\item{block.column}{Character string.  Column in sample metadata file
that contains blocking factors.}

\item{contrast.levels}{Character vector.  The contrast levels listed
in the order they should be compared as present in the contrast
column of the metadata file (e.g. c("WT", "KO")).}
}
\value{
A dataframe of the sample metadata.
}
\description{
Loads a tab-delimited file containing the sample metadata.  There should be
at least two columns - one called basename (containing the basename of the
samples) and another containing the contrasts of interest.  Another column
containing an additional factor of interest which should be taken into
account may also be included.
}
