% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bigwig.R
\name{make_bigwig}
\alias{make_bigwig}
\title{Create bigwig files from BAM}
\usage{
make_bigwig(bamfile = NULL, scale = 2e+07)
}
\arguments{
\item{bamfile}{Path to a BAM file.}

\item{scale}{The number of reads to scale to.  If NULL no scaling performed.}
}
\value{
Coverage file in bigwig format.
}
\description{
Script to generate bigwig files with the option for
normalisation from a BAM file.  Currently will only work with paired end
BAM files.
}
\examples{
\dontrun{
make_bigwig(bamfile = "HB1_sample.bam", scale = 20000000)
}

}
