% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_venn.R
\name{make_venn}
\alias{make_venn}
\title{Make venn diagrams showing overlapping gene lists}
\usage{
make_venn(results = results, p.value = 0.05)
}
\arguments{
\item{results}{Object produced by \code{run_dealr()} function.}

\item{p.value}{p-value for determining signficance for differential
expression.}
}
\value{
Venn diagrams showing overlap of gene lists between edgeR, limma, and DESeq2
}
\description{
Make venn diagrams showing overlapping gene lists
}
