% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fastqc.R
\name{plot_fqc}
\alias{plot_fqc}
\title{Plot FastQC Results}
\usage{
plot_fqc(fqcRes, modules = "all")
}
\arguments{
\item{fqcRes}{An object of class fqcRes or a FASTQC zipped result file.}

\item{modules}{Character vector containing the names of fastqc modules for
  which you want to import the data. Default is all. Allowed values include
  one or the combination of:
  \itemize{
  \item "summary",
  \item "basic_statistics",
  \item "per_base_sequence_quality",
  \item "per_sequence_quality_scores",
  \item "per_base_sequence_content",
  \item "per_sequence_gc_content",
  \item "per_base_n_content",
  \item "sequence_duplication_levels",
  \item "overrepresented_sequences",
  \item "adapter_content",
  \item "kmer_content"
  }
Partial match of module names are allowed and case is ignored.  Also, white
space may be used in place of underscore. For example, you can use
modules = "GC content", instead of "per_sequence_gc_content").}
}
\value{
Returns a list of ggplots (+/- dataframes) containing plots of the
  specified modules.
}
\description{
Plot FastQC data
}
