% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_counts.R
\name{process_counts}
\alias{process_counts}
\title{Pre-processing of counts}
\usage{
process_counts(se = NULL, cpm.cutoff = 1)
}
\arguments{
\item{se}{SummarizedExperiment object containing count matrix.}

\item{cpm.cutoff}{The counts-per-million threshold for filtering counts.}
}
\value{
SummarizedExperiment object containing the raw and filtered counts.
}
\description{
Function to proces raw counts before differential analysis.
Removes counts below the counts-per-million threshold set by the user and
generates diagnostic plots showing the count distribution before and after.
}
\examples{
\dontrun{
filtered_se <- process_counts(se = se, cpm.cutoff = 1)
}
}
