% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastqc.R
\name{read_fastqc}
\alias{read_fastqc}
\title{Read a single FastQC data file}
\usage{
read_fastqc(fqc)
}
\arguments{
\item{fqc}{Path to the FastQC result zip file to be imported.}
}
\value{
Returns an object of class fqRes containing the data from each of
 the FastQC modules.
}
\description{
Read FastQC data into R
}
\examples{
\dontrun{
read_fastqc("qcfile.zip")
}

}
