% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multifastqc.R
\name{read_multifastqc}
\alias{read_multifastqc}
\title{Read a a directory of FastQC data files}
\usage{
read_multifastqc(fqc.dir)
}
\arguments{
\item{fqc.dir}{Path to the directory of FASTQC zip files to be imported.}
}
\value{
Returns an object of class fqRes_multi.  Each object within the list
 is a fqRes object containing the FastQC data for a specific sample.
}
\description{
Read a directory of FastQC data files into R
}
\examples{
\dontrun{
read_multifastqc("qcfile.zip")
}

}
