% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diff.R
\name{run_diff}
\alias{run_diff}
\title{Run the differential analysis pipeline}
\usage{
run_diff(threads = 1, experimentTitle = "", modules = "LED",
  p.value = 0.05, sample.path = NULL, contrast.column = NULL,
  block.column = NULL, contrast.levels = NULL, count = TRUE,
  count.file = NULL, featurecounts = "featureCounts", bamfiles = NULL,
  annotationFile = NULL, annotationFormat = NULL,
  requireBothEndsMapped = TRUE, excludeChimeric = TRUE, pairedEnd = TRUE,
  countMultiMapping = FALSE, multiFeatureReads = TRUE, ignoreDup = FALSE,
  outname = "counts.txt", cpm.cutoff = 1, design = NULL,
  contrast = NULL, norm.method = "TMM", adjust.method = "BH")
}
\arguments{
\item{threads}{Number of threads to utilise where parallel
processing is possible.}

\item{experimentTitle}{Name of the experiment.}

\item{modules}{The differential expression analysis method
to be used. May be any combination of 'L' (limma), 'E' (edgeR), and 'D'
(DESeq2).}

\item{p.value}{Value between 0 and 1 specifying the adjusted
p.value threshold for significance.}

\item{sample.path}{Full path to the sample metadata file.}

\item{contrast.column}{Column in sample metadata file containing
information about the contrasts of interest.}

\item{block.column}{Column in sample metadata file containing
information about blocking factors or batch effect.}

\item{contrast.levels}{The order in which the contrasts should be evaluated.}

\item{count}{Boolean (TRUE or FALSE) indicating whether to perform counting.}

\item{count.file}{The path to the file containing the count matrix
(if count = F).  The first column should be the gene/feature names and the
subsequent columns the counts of reads in each sample.  The sample column
names must match the names in the sample metadata table.}

\item{featurecounts}{The path to featureCounts (if not in executable path).}

\item{bamfiles}{Vector of full path to bam files to be counted (if count = T).}

\item{annotationFile}{Path to annotation file.  If NULL, the default hg19
genome is used.}

\item{annotationFormat}{Specify the format of the annotation file. Acceptable
formats include ‘GTF’ and ‘SAF’.  The in-built annotation is 'SAF'.}

\item{requireBothEndsMapped}{Logical.  If TRUE, only fragments that have both
ends successfully aligned will be considered for summarization. This option
should be used together with pairedEnd = TRUE.}

\item{excludeChimeric}{Logical.  If TRUE, the chimeric fragments (those
fragments that have their two ends aligned to different chromosomes) will
NOT be counted. This option should be used together with pairedEnd = TRUE.}

\item{pairedEnd}{Logical. If TRUE,  fragments (or templates) will be counted
instead of reads. This option is only applicable for paired-end reads.}

\item{countMultiMapping}{If TRUE, multi-mapping reads/fragments will be
counted.}

\item{multiFeatureReads}{Reads/fragments overlapping with more than one
meta-feature/feature will be counted more than once. Note that when
performing meta-feature level summarization, a read (or fragment) will
still be counted once if it overlaps with multiple features within the same
meta-feature (as long as it does not overlap with other metafeatures).}

\item{ignoreDup}{Logical.  If TRUE, reads that were marked as duplicates will
be ignored.  In paired end data, the entire read pair will be ignored if at
least one end is found to be a duplicate read.}

\item{outname}{Character string.  Name of the output file. The output file
contains the number of reads assigned to each meta-feature or feature.}

\item{cpm.cutoff}{The counts-per-million threshold for filtering counts.}

\item{design}{Custom design matrix that may be used in place of default
matrix created by pipeline.  Only used for edgeR and limma analyses.}

\item{contrast}{Custom contrast matrix that may be used in place of default
matrix created by pipeline.  Only used for edgeR and limma analyses.}

\item{norm.method}{Normalisation method used by edgeR/limma analyses.  May
be "TMM", "RLE", "upperquartile" or "none".}

\item{adjust.method}{Method to be used for adjustment of nominal p-values.
May be one of "BH", "bonferroni", "holm", "hochberg", "hommel", "BY".}
}
\value{
Results of differential analysis.
}
\description{
Function to run the differential analysis portion of the pipeline.  There are
two possible starting points:
\itemize{
  \item{BAM files}{If starting with aligned BAM files, then set count = T.
        The pipeline will then begin with read counting before progressing to
        differential analysis.}
  \item{Count files}{If starting with count files generated externally, then
        set count = F and specify the text file containing the matrix of
        counts.  The pipeline will proceed directly to the differential
        analysis. The row names of the matrix should indicate the gene and
        the column name the library.  It is important that column names match
        the baseline in the sample metadata file.}
}
}
