% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastqc.R
\name{run_fastqc}
\alias{run_fastqc}
\title{Run FastQC}
\usage{
run_fastqc(fq.files = NULL, out.dir = NULL, threads = 1,
  fastqc = "fastqc")
}
\arguments{
\item{fq.files}{Vector of the full names and paths of the fastq files.}

\item{out.dir}{Path to the directory to which to write the results.  If NULL,
which is the default, a directory named "FASTQC" is created in the current
working directory.}

\item{threads}{The number of threads to use.  The default is 1.}

\item{fastqc}{The path to the fastqc executable.  On Unix systems, if the
executable is in $PATH, then it may be left as the default ("fastqc").
If it is not in $PATH, then the absolute path should be given.  If using
the WSL on Windows 10, then the path must be the absolute path in WSL.}
}
\value{
A directory containing the QC reports
}
\description{
Run the FastQC tool
}
\examples{
\dontrun{
fq.files <- list.files(path = "data-raw", pattern = "*.fastq$")
fastqc(fq.files, out.dir = "FASTQ", threads = 2, fastqc = "~/bin/fastqc")
}

}
