% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_goseq.R
\name{run_goseq}
\alias{run_goseq}
\alias{run_goseqmsigdb}
\title{Gene ontology and pathway analysis of DE genes}
\usage{
run_goseq(universe = universe, degs = degs, test.cats = c("GO:BP",
  "GO:MF"), numInCat = 1000, qv = 0.05, out = ".")

run_goseqmsigdb(universe = universe, degs = degs, numInCat = 1000,
  qv = 0.05, out = "", symbol = "HGNC")
}
\arguments{
\item{universe}{List of all genes tested for differential expression.}

\item{degs}{List of differentially expressed genes.}

\item{test.cats}{GO categories to test [default = c("GO:BP", "GO:MF")]}

\item{numInCat}{Maximum number of genes in a GO term.}

\item{qv}{Threshold q-value for significance.}

\item{out}{Output directory.}

\item{symbol}{Character specifying whether the gene identifiers are hgnc gene
symbols ("HGNC") or entrezid ("ENTREZ").}
}
\value{
\code{run_goseq} returns a dataframe and tab-delimited file of significant
GO terms.

\code{run_goseqmsigdb} returns a dataframe and tab-delimited file of
significant terms from the Broad Institute's Molecular Signatures Database.
}
\description{
\code{run_goseq} is a function to perform gene ontology analysis
using the goseq tool.

\code{run_goseqmdb} is a function to perform pathway analysis
  using the goseq tool.
}
\examples{
\dontrun{
universe <- unlist(as.list(rownames(counts)))
de.genes <- er$results$`OX-BUFF`$genes
gor <- run_goseq(universe, degs = de.genes)

gomdbr <- run_goseqmsigdb(universe, degs = de.genes, symbol = "HGNC")
}
}
