% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_homer.R
\name{run_homer}
\alias{run_homer}
\title{Run HOMER for promoter motif enrichment}
\usage{
run_homer(genelist, genome = "human", output_dir, len = NULL, bg = NULL,
  start = NULL, end = NULL, nomask = FALSE, S = NULL, mis = NULL,
  noconvert = FALSE, norevopp = FALSE, nomotif = FALSE, basic = FALSE,
  bits = FALSE, nocheck = FALSE, noknown = FALSE, b = FALSE,
  nogo = TRUE, humanGO = TRUE, noweight = FALSE, noredun = FALSE,
  g = FALSE, cpg = FALSE, rand = FALSE, fdr = NULL, p = NULL)
}
\arguments{
\item{genelist}{Vector of the genes in which to look for enrichment.}

\item{genome}{The reference genome [default = "human"].}

\item{output_dir}{The directory in which to output the results.}

\item{len}{Length of motifs to be found  [default = c(8, 10, 12)].}

\item{bg}{File of gene IDs to use as background.  By default HOMER uses all other
promoters as the background set.}

\item{start}{Number of base pairs upstream of the promoter start site to begin
looking [default = -300].}

\item{end}{Number of base pairs downstream of the promoter start site to stop
looking [default = +50].}

\item{nomask}{Use unmasked version of the genome [default = FALSE].}

\item{S}{Number of motifs of each length to find [default = 25].}

\item{mis}{Number of mismatches allowed [default = 2].}

\item{noconvert}{Do not convert input files into unigene ids [default = FALSE].}

\item{norevopp}{Do not search the revese strand for motifs [default = FALSE].}

\item{nomotif}{Do not search for de novo motifs [default = FALSE].}

\item{basic}{Do not check de novo motifs dor similarity to known [default = FALSE].}

\item{bits}{Scale sequence logos by information content [default = FALSE].}

\item{nocheck}{Do not check for similarity between de novo and known
motifs [default = FALSE].}

\item{noknown}{Do not check for known motif enrichment [default = FALSE].}

\item{b}{Use binomial distribution to calculate p-values [default = FALSE;
uses hypergeometric instead].}

\item{nogo}{Do not search for GO enrichment [default = TRUE].}

\item{humanGO}{Convert IDs to human for GO analysis [default = TRUE].}

\item{noweight}{Do not perform GC correction [default = FALSE].}

\item{noredun}{Do not remove predetermined redundant
promoters/sequences [default = FALSE].}

\item{g}{Input file is a group file, i.e. 1st column = id,
2nd = 0 or 1 [1=target,0=back] [default = FALSE].}

\item{cpg}{Use CpG\% instead of GC\% for sequence normalization [default = FALSE].}

\item{rand}{randomise labels for target and backgound sequences [default = FALSE].}

\item{fdr}{Numerical indicating number of randomisations to perform for
empirical FDR calculation for de novo discovery.
[default = NULL; no fdr calculation]}

\item{p}{Number of processors to use [default = 1]}
}
\value{
Output from HOMER, and also an R list object of results.
}
\description{
Function to run the HOMER tool for motif finding.  This function searches
motifs enriched in the promoters of the gene list provided.
}
\examples{
\dontrun{
motif_results <- homer_motifs(de_genes, genome = "human", output_dir = "motifs")
}
}
