% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma.R
\name{run_limma}
\alias{run_limma}
\title{Run limma+voom}
\usage{
run_limma(dge = dge, design = design, contrast = contrast,
  adjust.method = "BH", p.value = 0.05)
}
\arguments{
\item{dge}{DGE object produced by \code{\link[edgeR]{DGEList}}.}

\item{design}{Design matrix produced by \code{create_design}.}

\item{contrast}{Contrast matrix produced by \code{create_contrast}.}

\item{adjust.method}{Method to be used for adjustment of nominal p-values.
May be one of "BH", "bonferroni", "holm", "hochberg", "hommel", "BY".}

\item{p.value}{Value between 0 and 1.  Adjusted p-value for the differential
expression analysis.}
}
\value{
Results of limma analysis
}
\description{
Function to run voom normalisation (with quality weights)
followed by limma on a DGE object.
}
\examples{
\dontrun{
qlf <- run_limma(dge, design, contrast.matrix)
}
}
