% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_samtools.R
\name{run_samsort}
\alias{run_samsort}
\alias{run_samindex}
\alias{run_samview}
\alias{run_samflagstat}
\title{Wrapper scripts for SAMtools functions}
\usage{
run_samsort(file = NULL, samtools = "samtools", outformat = "BAM",
  threads = 1, memory = "768M", sortbyname = FALSE, suffix = "",
  keep = TRUE)

run_samindex(samtools = "samtools", bamfile = NULL, threads = 1)

run_samview(samtools = "samtools", file = NULL, regions = NULL,
  chrom.sizes = NULL, include.flag = NULL, exclude.flag = NULL,
  minQual = NULL, outformat = NULL, outname = NULL,
  include.header = FALSE, count = FALSE, threads = 1, subsample = NULL,
  keep.paired = TRUE, keep.proper.pair = TRUE, remove.unmapped = FALSE,
  remove.not.primary = FALSE, remove.duplicates = FALSE,
  remove.supplementary.alignment = FALSE, remove.mitochondrial = NULL)

run_samflagstat(samtools = "samtools", threads = 1, bamfile = NULL)
}
\arguments{
\item{file}{A vector of characters specifying the path to the bam files.}

\item{samtools}{The path to samtools (if not in executable path).}

\item{outformat}{String specifying output format: ('SAM'/'BAM'/'CRAM').  For
\code{run_samview} this should be SAM or 'BAM' only.}

\item{threads}{A positive integer specifying the number of sorting and
compression threads}

\item{memory}{String specifying maximum memory per thread; suffix K/M/G
recognized.}

\item{sortbyname}{Boolean. If TRUE, reads are sorted by name. If FALSE,
reads are sorted by chromosome/position.}

\item{suffix}{Suffix to add to the basename of the file e.g, _psort (for
position-sorted).}

\item{keep}{Boolean.  If TRUE, the input file is kept.  If FALSE, the input
file is deleted after a successful sort.}

\item{bamfile}{Vector of characters specifying the path to sorted BAM files.}

\item{regions}{Either path to a single BED file containing regions of
interest or regions defined in the format "RNAME:START[-END]".}

\item{chrom.sizes}{Path to chromosome size file. Required if converting to
SAM format and including header.  If not available, then one will be
generated automatically from the existing header.  This is a tab-delimited
text file where each line contains the reference name in the first column
and length in the second column.}

\item{include.flag}{Integer value.  Include output of alignments with any
bits set in this value present in the FLAG field.  See
\url{https:://broadinstitute.github.io}.}

\item{exclude.flag}{Integer value.  Exclude output of alignments with any
bits set in this value present in the FLAG field.  See
\url{https:://broadinstitute.github.io}.}

\item{minQual}{Skip alignments with MAPQ smaller than this value.}

\item{outname}{Name for output file.}

\item{include.header}{Boolean.  If TRUE header is included.  This option is
only necessary if outformat = "SAM" as header always included in "BAM".}

\item{count}{Boolean.  If true, only count alignments instead of printing.}

\item{subsample}{Float value - if set then subsampling is performed.  The
integer part is used to seed the random number generator and the part after
the decimal sets the fraction of reads to subsample.}

\item{keep.paired}{Boolean.  If TRUE, keep paired reads only.}

\item{keep.proper.pair}{Boolean.  If TRUE, keep concordantly paired reads
only.}

\item{remove.unmapped}{Boolean.  If TRUE, remove unmapped reads.}

\item{remove.not.primary}{Boolean.  If TRUE, remove reads mapped as secondary
alignments.}

\item{remove.duplicates}{Boolean.  If TRUE, remove reads marked as optical or
PCR duplicates.}

\item{remove.supplementary.alignment}{Boolean.  If TRUE, remove reads marked
as supplementary alignment.}

\item{remove.mitochondrial}{Character string.  If set, this will remove reads
mapping to the mitochondrial genome.  The string should match the reference
name for the mitochindrial genome in the alignment file.  Examples include
"ChrM", "M" and "MT".}
}
\value{
\code{run_samsort} returns a sorted alignment file.

\code{run_samindex} returns a BAI-format index file for sorted
BAM files.

\code{run_samview} returns aligned reads according to the regions and
filters specified in BAM or SAM format.

\code{run_samflagstat} returns a dataframe of alignment statistics.
}
\description{
\code{run_samsort} sorts alignment (SAM/BAM/CRAM) files either by position
or read name using \code{samtools sort}.

\code{run_samindex} indexes sorted BAM files using
\code{samtools index}.

\code{run_samview} outputs all alignments matching the flag and
  region filters specified in either SAM or BAM format using
  \code{samtools view}.

\code{run_samflagstat} uses \code{samtools flagstat} to
calculate and print statistics from a BAM file.  It provides counts for 13
categories of reads: total, secondary, supplementary, duplicates, mapped,
paired in sequencing, read1, read2, properly paired, with itself and its mate
mapped, singletons, mate mapped to a different chromosome, with mate mapped
to a different chromosome with MAPQ>5.
}
\examples{
\dontrun{
run_samsort(file = "HB1_sample.sam", samtools = "samtools",
            outformat = "BAM", threads = (parallel::detectCores() - 1),
            memory = "1G", sortbyname = FALSE, suffix = "", keep = TRUE)

}

\dontrun{
run_samindex(samtools = "samtools", bamfile = "HB1_sample.bam",
threads = (parallel::detectCores() - 1))
}

\dontrun{
run_samview(samtools = "samtools", file = "HB1_sample.bam", regions = NULL,
chrom.sizes = NULL, include.flag = NULL, exclude.flag = NULL,
minQual = NULL, outformat = "BAM", outname = NULL, include.header = FALSE,
count = FALSE, threads = (parallel::detectCores() - 1), subsample = NULL,
keep.paired = TRUE, keep.proper.pair = TRUE, remove.unmapped = TRUE,
remove.not.primary = TRUE, remove.duplicates = TRUE,
remove.supplementary.alignment = TRUE, remove.mitochondrial = "ChrM")
}

\dontrun{
run_samflagstat(samtools = "samtools", threads = parallel::detectCores(),
                bamfile = "HB1_sample.bam")
}

}
