% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelldist.R
\name{spelldist_old}
\alias{spelldist_old}
\title{Spelling distance matrix}
\usage{
spelldist_old(real, ideal = NULL, spelling_settings = NULL)
}
\arguments{
\item{real}{a vector of word forms as the scribe wrote them, case-insensitive.}

\item{ideal}{a vector of word forms in correct spelling, case-insensitive. If NULL, will use \code{real}, thus returning a symmetric distance matrix.}

\item{spelling_settings}{a list of spelling settings, such as spelling weights (=penalties). If NULL, will use the default settings.}
}
\value{
a matrix of spelling distances.
}
\description{
Calculate spelling distance matrix according to weights of insertion, deletion and substitution of characters of character combinations
}
\details{
Due to different spelling weights, the distance matrix may not by symmetric. Zero weights (penalties) for insertion or substitution do not require the spelling-distance algorithm (it is simply spelling normalization). Hence, setting some of these weights to zero improves the speed of calculation compared to however small but non-zero weights.
}
