% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelldist.R
\name{spelling_settings_default}
\alias{spelling_settings_default}
\title{Spelling settings}
\usage{
spelling_settings_default()
}
\value{
a list of spelling settings:
\item{characters}{Vector of characters. Some character combinations are represented as special characters, so they can be treated as a single character}
\item{combinations}{Vector of character combinations included in \code{characters}.}
\item{ins_weights}{A named vector of weights for inserting a character. E.g. ins_weights["a"] is the penalty for the scribe inserting an "a".}
\item{del_weights}{A named vector of weights for deleting (omitting) a character. E.g. del_weights["a"] is the penalty for the scribe omitting an "a".}
\item{sub_weights}{A named matrix of weights for substituting one character instead of another. E.g. sub_weights["a","b"] is the penalty for the scribe writing "a" instead of "b".}
}
\description{
Generate the default spelling error weights and some other spelling settings
}
