\name{grammar.symbols}
\Rdversion{1.1}
\alias{grammar.symbols}
\title{
	Read the grammar.output file generated by bison
}
\description{
	When bison is used with the \code{-v} switch, the 
	\code{grammar.output} file is generated. This function extracts
	symbols, terminal and non-terminal, from the file
	and structure them in a data frame
}
\usage{
grammar.symbols()
}
\value{
	A data frame. See \link{symbols}
}
\author{
	Romain Francois <francoisromain@free.fr>
}
\seealso{
	The data frame \code{\link{symbols}} is generated using this function
}
\examples{
\dontrun{
# the bison generated grammar output
system.file( "grammar", "gram.output", package = "parser" )
data( grammar_symbols )
% test if the grammar_symbols.txt is up to date
stopifnot( identical( grammar_symbols, grammar.symbols( ) ) )
}
}
\keyword{ manip }

