% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelldist.R
\name{spelldist}
\alias{spelldist}
\title{Spelling distance matrix}
\usage{
spelldist(real, ideal = NULL, spelling_settings = NULL)
}
\arguments{
\item{real}{a vector of word forms as the scribe wrote them, case-insensitive.}

\item{ideal}{a vector of word forms in correct spelling, case-insensitive. If NULL, will use \code{real}, thus returning a square distance matrix.}

\item{spelling_settings}{a list of spelling error weights. If NULL, will use the default settings.}
}
\value{
a matrix of spelling distances.
}
\description{
Calculate spelling distance matrix according to weights of insertion, deletion and substitution of characters of character combinations
}
